rename tasks, service, view, serializer, and model module files for better logging
