+++ server/apps/trak/admin.py
@@ -5,8 +5,8 @@ from .models import (
     Contact,
     EpaPhone,
     Handler,
-    ManifestHandler,
     Manifest,
+    ManifestHandler,
     RcraProfile,
     Site,
     SitePermission,
