+++ server/apps/trak/serializers/__init__.py
@@ -1,12 +1,13 @@
-from .address import AddressSerializer
-from .contact import ContactSerializer, EpaPhoneSerializer
-from .handler import HandlerSerializer
-from .manifest import ManifestSerializer
-from .rcra_profile import (
+from .address_ser import AddressSerializer
+from .contact_ser import ContactSerializer, EpaPhoneSerializer
+from .handler_ser import HandlerSerializer, ManifestHandlerSerializer
+from .manifest_ser import ManifestSerializer, MtnSerializer
+from .rcra_profile_ser import (
     EpaPermissionSerializer,
+    ProfileGetSerializer,
     ProfileUpdateSerializer,
     SitePermissionSerializer,
 )
-from .sites import SiteSerializer
-from .transporter import TransporterSerializer
-from .waste_line import WasteLineSerializer
+from .sites_ser import SiteSerializer
+from .transporter_ser import TransporterSerializer
+from .waste_line_ser import WasteLineSerializer
