+++ server/apps/trak/serializers/transporter_ser.py
@@ -1,6 +1,6 @@
 from apps.trak.models import Transporter
-from apps.trak.serializers.handler import HandlerSerializer
-from apps.trak.serializers.trak import TrakBaseSerializer
+from apps.trak.serializers.handler_ser import HandlerSerializer
+from apps.trak.serializers.trak_ser import TrakBaseSerializer
 
 
 class TransporterSerializer(TrakBaseSerializer):
