+++ server/apps/trak/services/__init__.py
@@ -1,4 +1,5 @@
-from .handler import HandlerService
-from .manifest import ManifestService
-from .rcrainfo import RcrainfoService
-from .site import SiteService
+from .handler_service import HandlerService
+from .manifest_service import ManifestService
+from .profile_service import RcraProfileService
+from .rcrainfo_service import RcrainfoService
+from .site_service import SiteService
