+++ server/apps/trak/services/handler_service.py
@@ -8,7 +8,7 @@ from rest_framework.exceptions import ValidationError
 from apps.trak.models import Handler
 from apps.trak.serializers import HandlerSerializer
 
-from .rcrainfo import RcrainfoService
+from .rcrainfo_service import RcrainfoService
 
 
 class HandlerService:
