+++ server/apps/trak/tasks/__init__.py
@@ -1,6 +1,6 @@
-from .handler_tasks import get_handler
-from .lookup_tasks import pull_federal_codes
-from .manifest_tasks import pull_manifest
-from .poc_tasks import say_hello
-from .profile_tasks import sync_user_sites
-from .site_tasks import sync_site_manifests
+from .handler_task import get_handler
+from .lookup_task import pull_federal_codes
+from .manifest_task import pull_manifest
+from .poc_task import say_hello
+from .profile_task import sync_user_sites
+from .site_task import sync_site_manifests
