+++ server/apps/trak/tasks/profile_task.py
@@ -4,7 +4,7 @@ from celery import Task, shared_task, states
 from celery.exceptions import Ignore, Reject
 from requests import RequestException
 
-from apps.trak.services.profile import RcraProfileService
+from apps.trak.services import RcraProfileService
 
 logger = logging.getLogger(__name__)
