+++ server/apps/trak/views/__init__.py
@@ -1,4 +1,4 @@
-from .handler import HandlerSearch, HandlerView, TransporterView
-from .manifest import ManifestView, MtnList, PullManifest
-from .rcra_profile import RcraProfileView, SitePermissionView, SyncProfile
-from .site import SiteApi, SiteList, SiteManifest, SyncSiteManifest
+from .handler_view import HandlerSearch, HandlerView, ManifestHandlerView, TransporterView
+from .manifest_view import ManifestView, MtnList, PullManifest
+from .rcra_profile_view import RcraProfileView, SitePermissionView, SyncProfile
+from .site_view import SiteApi, SiteList, SiteManifest, SyncSiteManifest
